ObjCompiler: compile to new format with .mdl and .mdg files
Update ObjCompiler.java to compile .obj model files to this new format,
where the model data is split across a .mdl and a .mdg file. This allows
data intended for CPU vs GPU memory to be separated.

In addition, the format allows for storing multiple submeshes in a
single model asset.